% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{circle}
\alias{circle}
\title{Plot a circle
Plot a circle}
\usage{
circle(x, y, r, from = 0, to = 2 * pi, lines = FALSE, ...)
}
\arguments{
\item{x}{x-axis position of circle.}

\item{y}{y-axis position of circle.}

\item{r}{Radius of circle.}

\item{from}{Measurment in radians.}

\item{to}{Measurement in radians.}

\item{lines}{TBD}

\item{...}{Additional arguments to \code{\link[graphics]{lines}}.}
}
\value{
Lines of a parametrized circle.
}
\description{
Plot a circle
Plot a circle
}
