% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_colors.R
\name{color_dist}
\alias{color_dist}
\title{Find Colors with Greatest Distance
Find the hex color in the color vector \code{b} that has the greatest (Euclidean) distance from the color \code{a}.}
\usage{
color_dist(a, b, omit = NULL)
}
\arguments{
\item{a}{Single color as a hex value.}

\item{b}{Color vector as hex values.}

\item{omit}{Color vector as hex values to omit from distance calculation.}
}
\value{
The hex color from \code{b} that has the largest distance from \code{a}
}
\description{
Find Colors with Greatest Distance
Find the hex color in the color vector \code{b} that has the greatest (Euclidean) distance from the color \code{a}.
}
