% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gif.R
\name{create_gif_frame}
\alias{create_gif_frame}
\title{Create a gif of an event}
\usage{
create_gif_frame(event, back, loop_iter, static = FALSE, legend = TRUE)
}
\arguments{
\item{event}{An integer that identifies the event as a row in the data frame.}

\item{back}{An integer representing the number of time steps prior to the event.}

\item{N}{An integer representing the number of iterations (moving forward in time).}

\item{fname}{A string ending in \code{.gif}.}
}
\value{
A gif of the basketball play.
}
\description{
Create a gif of an event
}
