% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_shots.R
\name{extract_shots}
\alias{extract_shots}
\title{Extract Shots from SportVU Data}
\usage{
extract_shots(obj, pbp)
}
\arguments{
\item{obj}{A data frame returned from \code{\link[NBAsportvu]{sportvu_df}}.}

\item{pbp}{Play-by_play data frame (e.g. an object returned by
\code{\link[NBAapi]{get_pbp}}). One of the columns must contain the
\code{PCTIMESTRING} variable.}
}
\description{
Extract all field goals from the SportVU data frame within a certain time margin of the field goal occuring.
}
