% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{fullcourt_lines}
\alias{fullcourt_lines}
\title{Plot Fullcourt
Plot an NBA regulation full basketball court.}
\usage{
fullcourt_lines()
}
\value{
Lines corresponding to a full basketball court.
}
\description{
Plot Fullcourt
Plot an NBA regulation full basketball court.
}
