% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{jersey}
\alias{jersey}
\title{Jersey Number For internal use. Get jersey number from player ID.}
\usage{
jersey(playerid, player_table)
}
\arguments{
\item{playerid}{Player ID number as a number.}

\item{player_table}{Table from SportVU indicating the player's name, jersey
number, ID, position, and team affiliation.}
}
\value{
Jersey number that corresponds to player ID.
}
\description{
Jersey Number For internal use. Get jersey number from player ID.
}
