% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fullcourt.R
\name{plot_fullcourt}
\alias{plot_fullcourt}
\title{Plot Full Court}
\usage{
plot_fullcourt()
}
\value{
A plot of the NBA regulation court.
}
\description{
Plot the NBA regulation court. Wrapper for
\code{\link[NBAsportvu]{fullcourt_lines}}.
}
\examples{
\dontrun{
plot_fullcourt()
}

}
