% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gif.R
\name{plot_shot}
\alias{plot_shot}
\title{Plot an Event.}
\usage{
plot_shot(obj, loop = 1, static = FALSE)
}
\arguments{
\item{obj}{A subset of the data frame object returned from the
\code{\link[NBAsportvu]{sportvu_df}} function.}

\item{loop}{Integer used to loop over each time point of \code{obj} when
using the \code{\link[animation]{saveGIF}} function.}

\item{static}{Logical variable which defaults to \code{FALSE}. If looping
over \code{obj} for use with \code{\link[animation]{saveGIF}} then this
should be set to \code{TRUE}.}
}
\value{
A plot of a basketball event.
}
\description{
Plot an Event.
}
