% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sportvu_df.R
\name{sportvu_df}
\alias{sportvu_df}
\title{Convert SportVU json to Data Frame.}
\usage{
sportvu_df(path)
}
\arguments{
\item{path}{Path to SportVU json file as a string.}
}
\value{
A data frame that includes motion tracking information for a game.
  Variables \code{quarter}, \code{game_clock}, and \code{shot_clock} are self
  explanatory. Variables \code{x} and \code{y} denotes the spatial location
  of the ball, and \code{z} denotes the height of the ball. For the remaining
  variables, the first letter in the variable name is either \code{a} for the
  away team or \code{h} for the home team. Variables \code{a*_team} and
  \code{a*_ent} are the team and player identifiers. The spatial coordinates
  of each player are denoted by \code{a*_x} and \code{a*_y}.
}
\description{
Convert SportVU json file of NBA player tracking data to a data frame. Note,
due to the size of the player tracking data a single game can take a few
minutes to convert.
}
\examples{
\dontrun{
game <- sportvu_df("~/path_to_file/0021500411.json")
}

}
