% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_colors.R
\name{team_col_map}
\alias{team_col_map}
\title{Team Color Map}
\usage{
team_col_map()
}
\value{
A list with length equal to the number of teams. Each team name
  element in the list corresponds to a vector of hex colors that represent
  the team's official colors.
}
\description{
Mapping of official team colors.
}
